Step Fourth Task1 Final

+++ src/com/course/syntax/StepFourthTask1.java
@@ -6,7 +6,7 @@ public class StepFourthTask1 {
     public static void main(String[] args) {
         Scanner scanner = new Scanner(System.in);
         System.out.println("Enter the number: ");
-        if (scanner.hasNextFloat()) {
+        if (scanner.hasNextInt()) {
             int number = scanner.nextInt();
             if (number < 0) {
                 System.out.println("Error! The number is negative");
